+++ tests/unit/models/profile-test.js
@@ -1,13 +1,13 @@
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 
-module('Unit | Model | perfil', function(hooks) {
+module('Unit | Model | profile', function(hooks) {
   setupTest(hooks);
 
   // Replace this with your real tests.
   test('it exists', function(assert) {
     let store = this.owner.lookup('service:store');
-    let model = store.createRecord('perfil', {});
+    let model = store.createRecord('profile', {});
     assert.ok(model);
   });
 });
